+++ colander/tests/test_colander.py
@@ -110,9 +110,10 @@ class TestInvalid(unittest.TestCase):
         exc1.add(exc2, 2)
         exc2.add(exc3, 3)
         d = exc1.asdict()
-        self.assertEqual(d,
-                         {'node1.node2.3': 'exc1; exc2; validator1; validator2',
-                          'node1.node3': 'exc1; message1'})
+        self.assertEqual(
+            d,
+            {'node1.node2.3': 'exc1; exc2; validator1; validator2',
+             'node1.node3': 'exc1; message1'})
 
     def test_asdict_with_all_validator_functional(self):
         # see https://github.com/Pylons/colander/issues/2
@@ -153,7 +154,8 @@ class TestInvalid(unittest.TestCase):
         result = str(exc1)
         self.assertEqual(
             result,
-            "{'node1.node2.3': 'exc1; exc2; exc3', 'node1.node4': 'exc1; exc4'}"
+            "{'node1.node2.3': 'exc1; exc2; exc3', "
+            "'node1.node4': 'exc1; exc4'}"
             )
 
     def test___setitem__fails(self):
@@ -425,15 +427,17 @@ class TestEmail(unittest.TestCase):
         validator = self._makeOne()
         from colander import Invalid
         self.assertRaises(Invalid, validator, None, 'me@here.')
-        self.assertRaises(Invalid, validator, None, 'name@here.tldiswaytoolooooooooong')
+        self.assertRaises(Invalid,
+                          validator, None, 'name@here.tldiswaytoolooooooooong')
         self.assertRaises(Invalid, validator, None, '@here.us')
         self.assertRaises(Invalid, validator, None, 'me@here..com')
         self.assertRaises(Invalid, validator, None, 'me@we-here-.com')
 
+
 class TestLength(unittest.TestCase):
-    def _makeOne(self, min=None, max=None):
+    def _makeOne(self, **kw):
         from colander import Length
-        return Length(min=min, max=max)
+        return Length(**kw)
 
     def test_success_no_bounds(self):
         validator = self._makeOne()
@@ -460,6 +464,17 @@ class TestLength(unittest.TestCase):
         validator = self._makeOne(max=1)
         e = invalid_exc(validator, None, 'ab')
         self.assertEqual(e.msg.interpolate(), 'Longer than maximum length 1')
+
+    def test_min_failure_msg_override(self):
+        validator = self._makeOne(min=1, min_err='Need at least ${min}, mate')
+        e = invalid_exc(validator, None, [])
+        self.assertEqual(e.msg.interpolate(), 'Need at least 1, mate')
+
+    def test_max_failure_msg_override(self):
+        validator = self._makeOne(max=1, max_err='No more than ${max}, mate')
+        e = invalid_exc(validator, None, [1, 2])
+        self.assertEqual(e.msg.interpolate(), 'No more than 1, mate')
+
 
 class TestOneOf(unittest.TestCase):
     def _makeOne(self, values):
@@ -535,6 +550,57 @@ class Test_url_validator(unittest.TestCase):
 
     def test_it_failure(self):
         val = 'not-a-url'
+        from colander import Invalid
+        self.assertRaises(Invalid, self._callFUT, val)
+
+class TestUUID(unittest.TestCase):
+    def _callFUT(self, val):
+        from colander import uuid
+        return uuid(None, val)
+
+    def test_success_hexadecimal(self):
+        val = '123e4567e89b12d3a456426655440000'
+        result = self._callFUT(val)
+        self.assertEqual(result, None)
+
+    def test_success_with_dashes(self):
+        val = '123e4567-e89b-12d3-a456-426655440000'
+        result = self._callFUT(val)
+        self.assertEqual(result, None)
+
+    def test_success_upper_case(self):
+        val = '123E4567-E89B-12D3-A456-426655440000'
+        result = self._callFUT(val)
+        self.assertEqual(result, None)
+
+    def test_success_with_braces(self):
+        val = '{123e4567-e89b-12d3-a456-426655440000}'
+        result = self._callFUT(val)
+        self.assertEqual(result, None)
+
+    def test_success_with_urn_ns(self):
+        val = 'urn:uuid:{123e4567-e89b-12d3-a456-426655440000}'
+        result = self._callFUT(val)
+        self.assertEqual(result, None)
+
+    def test_failure_random_string(self):
+        val = 'not-a-uuid'
+        from colander import Invalid
+        self.assertRaises(Invalid, self._callFUT, val)
+
+    def test_failure_not_hexadecimal(self):
+        val = '123zzzzz-uuuu-zzzz-uuuu-42665544zzzz'
+        from colander import Invalid
+        self.assertRaises(Invalid, self._callFUT, val)
+
+    def test_failure_invalid_length(self):
+        # Correct UUID: 8-4-4-4-12
+        val = '88888888-333-4444-333-cccccccccccc'
+        from colander import Invalid
+        self.assertRaises(Invalid, self._callFUT, val)
+
+    def test_failure_with_invalid_urn_ns(self):
+        val = 'urn:abcd:{123e4567-e89b-12d3-a456-426655440000}'
         from colander import Invalid
         self.assertRaises(Invalid, self._callFUT, val)
 
@@ -725,7 +791,7 @@ class TestMapping(unittest.TestCase):
         typ = self._makeOne()
         result = typ.serialize(node, {'a':drop})
         self.assertEqual(result, {})
-        
+
     def test_flatten(self):
         node = DummySchemaNode(None, name='node')
         int1 = DummyType()
@@ -1309,6 +1375,16 @@ class TestSequence(unittest.TestCase):
         result = typ.flatten(node, [1, 2])
         self.assertEqual(result, {'node.0': 1, 'node.1': 2})
 
+    def test_flatten_with_integer(self):
+        from colander import Integer
+        node = DummySchemaNode(None, name='node')
+        node.children = [
+            DummySchemaNode(Integer(), name='foo'),
+        ]
+        typ = self._makeOne()
+        result = typ.flatten(node, [1, 2])
+        self.assertEqual(result, {'node.0': 1, 'node.1': 2})
+
     def test_flatten_listitem(self):
         node = DummySchemaNode(None, name='node')
         node.children = [
@@ -1500,7 +1576,7 @@ class TestInteger(unittest.TestCase):
         result = typ.serialize(node, val)
         self.assertEqual(result, colander.null)
 
-    def test_serialize_emptystring(self):
+    def test_deserialize_emptystring(self):
         import colander
         val = ''
         node = DummySchemaNode(None)
@@ -2343,8 +2419,9 @@ class TestSchemaNode(unittest.TestCase):
 
     def test_ctor_no_title(self):
         child = DummySchemaNode(None, name='fred')
-        node = self._makeOne(None, child, validator=1, default=2, name='name_a',
-                             missing='missing')
+        node = self._makeOne(
+            None, child, validator=1, default=2,
+            name='name_a', missing='missing')
         self.assertEqual(node.typ, None)
         self.assertEqual(node.children, [child])
         self.assertEqual(node.validator, 1)
@@ -2463,6 +2540,19 @@ class TestSchemaNode(unittest.TestCase):
         e = invalid_exc(node.deserialize, 1)
         self.assertEqual(e.msg, 'Wrong')
 
+    def test_deserialize_with_unbound_validator(self):
+        from colander import Invalid
+        from colander import deferred
+        from colander import UnboundDeferredError
+        typ = DummyType()
+        def validator(node, kw):
+            def _validate(node, value):
+                node.raise_invalid('Invalid')
+            return _validate
+        node = self._makeOne(typ, validator=deferred(validator))
+        self.assertRaises(UnboundDeferredError, node.deserialize, None)
+        self.assertRaises(Invalid, node.bind(foo='foo').deserialize, None)
+
     def test_deserialize_value_is_null_no_missing(self):
         from colander import null
         from colander import Invalid
@@ -2483,6 +2573,14 @@ class TestSchemaNode(unittest.TestCase):
         node = self._makeOne(typ, missing_msg='Missing')
         e = invalid_exc(node.deserialize, null)
         self.assertEqual(e.msg, 'Missing')
+
+    def test_deserialize_value_with_interpolated_missing_msg(self):
+        from colander import null
+        typ = DummyType()
+        node = self._makeOne(typ, missing_msg='Missing attribute ${title}',
+                             name='name_a')
+        e = invalid_exc(node.deserialize, null)
+        self.assertEqual(e.msg.interpolate(), 'Missing attribute Name A')
 
     def test_deserialize_noargs_uses_default(self):
         typ = DummyType()
@@ -2726,6 +2824,22 @@ class TestSchemaNodeSubclassing(unittest.TestCase):
         node = MyNode()
         result = node.deserialize(colander.null)
         self.assertEqual(result, 10)
+
+    def test_subclass_uses_title(self):
+        import colander
+        class MyNode(colander.SchemaNode):
+            schema_type = colander.Int
+            title = 'some title'
+        node = MyNode(name='my')
+        self.assertEqual(node.title, 'some title')
+
+    def test_subclass_title_overwritten_by_constructor(self):
+        import colander
+        class MyNode(colander.SchemaNode):
+            schema_type = colander.Int
+            title = 'some title'
+        node = MyNode(name='my', title='other title')
+        self.assertEqual(node.title, 'other title')
 
     def test_subclass_value_overridden_by_constructor(self):
         import colander
@@ -3425,7 +3539,8 @@ class TestFunctional(object):
     def test_invalid_asdict(self):
         expected = {
             'schema.int': '20 is greater than maximum value 10',
-            'schema.ob': 'The dotted name "no.way.this.exists" cannot be imported',
+            'schema.ob': 'The dotted name "no.way.this.exists" '
+                         'cannot be imported',
             'schema.seq.0.0': '"q" is not a number',
             'schema.seq.1.0': '"w" is not a number',
             'schema.seq.2.0': '"e" is not a number',
@@ -3620,7 +3735,7 @@ class TestUltraDeclarative(unittest.TestCase, TestFunctional):
         return schema
 
 class TestDeclarativeWithInstantiate(unittest.TestCase, TestFunctional):
-    
+
     def _makeSchema(self, name='schema'):
 
         import colander
@@ -3634,20 +3749,20 @@ class TestDeclarativeWithInstantiate(unittest.TestCase, TestFunctional):
             ob = colander.SchemaNode(colander.GlobalObject(package=colander))
             @colander.instantiate()
             class seq(colander.SequenceSchema):
-                
+
                 @colander.instantiate()
                 class tup(colander.TupleSchema):
                     tupint = colander.SchemaNode(colander.Int())
                     tupstring = colander.SchemaNode(colander.String())
-                    
+
             @colander.instantiate()
             class tup(colander.TupleSchema):
                 tupint = colander.SchemaNode(colander.Int())
                 tupstring = colander.SchemaNode(colander.String())
-                
+
             @colander.instantiate()
             class seq2(colander.SequenceSchema):
-                
+
                 @colander.instantiate()
                 class mapping(colander.MappingSchema):
                     key = colander.SchemaNode(colander.Int())